# swagger.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/v1/departments":  {
                                              "post":  {
                                                           "operationId":  "DepartmentsController_create_v1",
                                                           "parameters":  [

                                                                          ],
                                                           "requestBody":  {
                                                                               "required":  true,
                                                                               "content":  {
                                                                                               "application/json":  {
                                                                                                                        "schema":  {
                                                                                                                                       "$ref":  "#/components/schemas/CreateDepartmentDto"
                                                                                                                                   }
                                                                                                                    }
                                                                                           }
                                                                           },
                                                           "responses":  {
                                                                             "201":  {
                                                                                         "description":  "The department has been successfully created."
                                                                                     }
                                                                         },
                                                           "security":  [
                                                                            {
                                                                                "access-token":  [

                                                                                                 ]
                                                                            }
                                                                        ],
                                                           "summary":  "Create a new department",
                                                           "tags":  [
                                                                        "Departments"
                                                                    ]
                                                       },
                                              "get":  {
                                                          "operationId":  "DepartmentsController_list_v1",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "page",
                                                                                 "required":  false,
                                                                                 "in":  "query",
                                                                                 "description":  "Page number",
                                                                                 "schema":  {
                                                                                                "minimum":  1,
                                                                                                "default":  1,
                                                                                                "type":  "number"
                                                                                            }
                                                                             },
                                                                             {
                                                                                 "name":  "limit",
                                                                                 "required":  false,
                                                                                 "in":  "query",
                                                                                 "description":  "Number of records per page",
                                                                                 "schema":  {
                                                                                                "minimum":  1,
                                                                                                "maximum":  100,
                                                                                                "default":  20,
                                                                                                "type":  "number"
                                                                                            }
                                                                             },
                                                                             {
                                                                                 "name":  "search",
                                                                                 "required":  false,
                                                                                 "in":  "query",
                                                                                 "description":  "Search by department name",
                                                                                 "schema":  {
                                                                                                "example":  "Eng",
                                                                                                "type":  "string"
                                                                                            }
                                                                             },
                                                                             {
                                                                                 "name":  "isActive",
                                                                                 "required":  false,
                                                                                 "in":  "query",
                                                                                 "description":  "Filter by active status",
                                                                                 "schema":  {
                                                                                                "example":  true,
                                                                                                "type":  "boolean"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "List of departments with pagination."
                                                                                    }
                                                                        },
                                                          "security":  [
                                                                           {
                                                                               "access-token":  [

                                                                                                ]
                                                                           }
                                                                       ],
                                                          "summary":  "List all departments",
                                                          "tags":  [
                                                                       "Departments"
                                                                   ]
                                                      }
                                          },
                  "/api/v1/departments/{id}":  {
                                                   "get":  {
                                                               "operationId":  "DepartmentsController_getById_v1",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "required":  true,
                                                                                      "in":  "path",
                                                                                      "description":  "Department ID",
                                                                                      "schema":  {
                                                                                                     "type":  "string"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "The department details."
                                                                                         }
                                                                             },
                                                               "security":  [
                                                                                {
                                                                                    "access-token":  [

                                                                                                     ]
                                                                                }
                                                                            ],
                                                               "summary":  "Get department by ID",
                                                               "tags":  [
                                                                            "Departments"
                                                                        ]
                                                           },
                                                   "patch":  {
                                                                 "operationId":  "DepartmentsController_update_v1",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "id",
                                                                                        "required":  true,
                                                                                        "in":  "path",
                                                                                        "description":  "Department ID",
                                                                                        "schema":  {
                                                                                                       "type":  "string"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "requestBody":  {
                                                                                     "required":  true,
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/UpdateDepartmentDto"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "The department has been successfully updated."
                                                                                           }
                                                                               },
                                                                 "security":  [
                                                                                  {
                                                                                      "access-token":  [

                                                                                                       ]
                                                                                  }
                                                                              ],
                                                                 "summary":  "Update department",
                                                                 "tags":  [
                                                                              "Departments"
                                                                          ]
                                                             },
                                                   "delete":  {
                                                                  "operationId":  "DepartmentsController_delete_v1",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "id",
                                                                                         "required":  true,
                                                                                         "in":  "path",
                                                                                         "description":  "Department ID",
                                                                                         "schema":  {
                                                                                                        "type":  "string"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "The department has been successfully deleted."
                                                                                            }
                                                                                },
                                                                  "security":  [
                                                                                   {
                                                                                       "access-token":  [

                                                                                                        ]
                                                                                   }
                                                                               ],
                                                                  "summary":  "Soft delete a department",
                                                                  "tags":  [
                                                                               "Departments"
                                                                           ]
                                                              }
                                               },
                  "/api/v1/departments/{id}/status":  {
                                                          "patch":  {
                                                                        "operationId":  "DepartmentsController_updateStatus_v1",
                                                                        "parameters":  [
                                                                                           {
                                                                                               "name":  "id",
                                                                                               "required":  true,
                                                                                               "in":  "path",
                                                                                               "description":  "Department ID",
                                                                                               "schema":  {
                                                                                                              "type":  "string"
                                                                                                          }
                                                                                           }
                                                                                       ],
                                                                        "requestBody":  {
                                                                                            "required":  true,
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/UpdateDepartmentStatusDto"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        }
                                                                                        },
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "The department status has been successfully updated."
                                                                                                  }
                                                                                      },
                                                                        "security":  [
                                                                                         {
                                                                                             "access-token":  [

                                                                                                              ]
                                                                                         }
                                                                                     ],
                                                                        "summary":  "Update department status (activate/deactivate)",
                                                                        "tags":  [
                                                                                     "Departments"
                                                                                 ]
                                                                    }
                                                      },
...
              },
    "info":  {
                 "title":  "CRM + HRMS API",
                 "description":  "Production API documentation for the CRM + HRMS platform.",
                 "version":  "1.0.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "access-token":  {
                                                                    "scheme":  "bearer",
                                                                    "bearerFormat":  "JWT",
                                                                    "type":  "http",
                                                                    "description":  "Enter JWT access token"
                                                                }
                                           },
                       "schemas":  {
...
                                       "CreateDepartmentDto":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "name":  {
                                                                                                   "type":  "string",
                                                                                                   "description":  "Name of the department",
                                                                                                   "example":  "Engineering"
                                                                                               },
                                                                                      "description":  {
                                                                                                          "type":  "string",
                                                                                                          "description":  "Description of the department",
                                                                                                          "example":  "Handles all technical development"
                                                                                                      }
                                                                                  },
                                                                   "required":  [
                                                                                    "name"
                                                                                ]
                                                               },
                                       "UpdateDepartmentDto":  {
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "name":  {
                                                                                                   "type":  "string",
                                                                                                   "description":  "Name of the department",
                                                                                                   "example":  "Engineering"
                                                                                               },
                                                                                      "description":  {
                                                                                                          "type":  "string",
                                                                                                          "description":  "Description of the department",
                                                                                                          "example":  "Handles all technical development"
                                                                                                      }
                                                                                  }
                                                               },
                                       "UpdateDepartmentStatusDto":  {
                                                                         "type":  "object",
                                                                         "properties":  {
                                                                                            "isActive":  {
                                                                                                             "type":  "boolean",
                                                                                                             "description":  "Active status of the department",
                                                                                                             "example":  true
                                                                                                         }
                                                                                        },
                                                                         "required":  [
                                                                                          "isActive"
                                                                                      ]
                                                                     },
...
                                   }
                   }
}